commit: working task3 M3

+++ WebApplication1/src/java/pack_M3/Saldo.java
@@ -11,6 +11,6 @@ package pack_M3;
  */
 public class Saldo {
     
-    private float saldo;
+    private double saldo;
     
 }
